% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{add_background}
\alias{add_background}
\title{Add background color}
\usage{
add_background(
  input,
  color = "black",
  rounding = 0,
  openfile = TRUE,
  output = input
)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{color}{Vector of background color}

\item{rounding}{Rounding of background rectangle's corners. The default is 0 (no rounding). A rounding value of 1 creates a circle. If rounding is a vector of 2 values, the x and y rounding parameters are set separately.}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Add background color
}
\examples{
\dontrun{
library(spiro)
spiro(
  fixed_radius = 3,
  cycling_radius = 1,
  file = "add_background.svg") \%>\%
  add_background(color = "black")
  }
}
