% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{add_background_gradient_radial}
\alias{add_background_gradient_radial}
\title{Add background color radial gradients}
\usage{
add_background_gradient_radial(input, colors = c("white", "black"),
  stops = seq(0, 1, length.out = length(colors)), center_x = 0.5,
  center_y = 0.5, focus_x = 0.5, focus_y = 0.5, radius = sqrt(2),
  rounding = 0, openfile = TRUE, output = input)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{colors}{Vector of gradient colors}

\item{stops}{Vector of stops for gradient colors. Ranges from 0 to 1.}

\item{center_x}{x-coordinate for center of gradient. Number between 0 and 1.}

\item{center_y}{y-coordinate for center of gradient. Number between 0 and 1.}

\item{focus_x}{x-coordinate for focus of gradient. Number between 0 and 1.}

\item{focus_y}{y-coordinate for focus of gradient. Number between 0 and 1.}

\item{radius}{Radius of radial gradient. The default is the square root of 2 so that the gradient extends to the corners of the square. If you want it to extend to the edges, set it to 1.}

\item{rounding}{Rounding of background rectangle's corners. The default is 0 (no rounding). A rounding value of 1 creates a circle. If rounding is a vector of 2 values, the x and y rounding parameters are set separately.}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Add background color radial gradients
}
\examples{
library(spiro)
library(magrittr)
spiro(fixed_radius = 3, cycling_radius = 1) \%>\%
    add_background_color(color = "black")
}
