% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{add_blur}
\alias{add_blur}
\title{Add blur}
\usage{
add_blur(
  input,
  stdDeviation = 1,
  tag = "g",
  feMerge = TRUE,
  offset_x = 0,
  offset_y = 0,
  filter_id = "blurMe",
  source = "SourceGraphic",
  openfile = TRUE,
  output = input
)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{stdDeviation}{Standard deviation of blur.}

\item{tag}{svg tag to which the blur should apply. Defaults to g.}

\item{feMerge}{Merge original graphic. Defaults to TRUE}

\item{offset_x}{offset of blue in x direction. Defaults to 0.}

\item{offset_y}{offset of blue in y direction. Defaults to 0.}

\item{filter_id}{Name of filter id. Defaults to blueMe}

\item{source}{Merge original image. Defaults to SourceGraphic. Options: SourceGraphic | SourceAlpha | BackgroundImage | BackgroundAlpha | FillPaint | StrokePaint}

\item{openfile}{Open file in default program for .svg format. Defaults to TRUE}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Add blur
}
