% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{add_circle}
\alias{add_circle}
\title{Add a circle to the image}
\usage{
add_circle(
  input,
  x = 0.5,
  y = 0.5,
  r = 1,
  color = "black",
  fill = "none",
  line_width = 1,
  openfile = TRUE,
  output = input
)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{x}{x-coordinate of circle center (0 to 1)}

\item{y}{y-coordinate of circle center (0 to 1)}

\item{r}{radius (0 to 1)}

\item{color}{line color}

\item{fill}{fill color}

\item{line_width}{width of line}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Add a circle to the image
}
\examples{
\dontrun{
spiro(fixed_radius = 3,
      cycling_radius = 1,
      file = "add_circle_example.svg",
      colors = "red") \%>\%
  add_circle(r = 0.308, fill = "white", color = "white")
  }
}
