% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{add_fills}
\alias{add_fills}
\title{Add fills to lines}
\usage{
add_fills(input, colors = "black", transparency = NA,
  rule = c("evenodd", "winding"), openfile = TRUE, output = input)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{colors}{Vector of background color}

\item{transparency}{Transparency of colors. Ranges from 0 to 1. Default is NA (i.e., leave the colors as specified in the colors argument).}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.
#' @return output name}

\item{line_width}{Width of lines. Default is 0.5.}
}
\description{
Add fills to lines
}
\examples{
library(spiro)
library(magrittr)
spiro(fixed_radius = 3, cycling_radius = 1) \%>\%
    add_background_color(color = "black")
}
