% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{add_lines}
\alias{add_lines}
\title{Add color to lines}
\usage{
add_lines(
  input,
  colors = "black",
  transparency = NA,
  line_width = 0.5,
  openfile = TRUE,
  output = input
)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{colors}{Vector of background color}

\item{transparency}{Transparency of colors. Ranges from 0 to 1. Default is NA (i.e., leave the colors as specified in the colors argument).}

\item{line_width}{Width of lines. Default is 0.5.}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.
#' @return output name}
}
\description{
Add color to lines
}
\examples{
\dontshow{.old_wd <- setwd(tempdir())}
library(spiro)
spiro(
  fixed_radius = 3,
  cycling_radius = 1,
  file = "add_lines.svg"
  ) \%>\%
  add_lines(
    colors = "firebrick",
    line_width = 3)
\dontshow{setwd(.old_wd)}
}
