% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{add_pathdot}
\alias{add_pathdot}
\title{Add a moving circle to paths. Resulting file best viewed in Google Chrome.}
\usage{
add_pathdot(
  input,
  colors = "red",
  transparency = NA,
  radius = 3,
  duration = 10,
  delay = 0,
  path_id = "pathdot",
  openfile = TRUE,
  output = input
)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{colors}{Vector of background color}

\item{transparency}{Transparency of colors. Ranges from 0 to 1. Default is NA (i.e., leave the colors as specified in the colors argument).}

\item{radius}{Radius of dot}

\item{duration}{Number seconds to complete animation}

\item{delay}{Number of seconds to delay animation}

\item{path_id}{Set id to path}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Add a moving circle to paths. Resulting file best viewed in Google Chrome.
}
\examples{
library(spiro)
spiro(
 fixed_radius = 3,
 cycling_radius = 1,
 file = "add_pathdot.svg"
 ) \%>\%
 add_pathdot(duration = 3)
}
