% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{add_protocol}
\alias{add_protocol}
\title{Add a test protocol to an exercise testing dataset}
\usage{
add_protocol(data, protocol)
}
\arguments{
\item{data}{A spiro \code{data.frame} containing the exercise testing data.}

\item{protocol}{A \code{data.frame} containing the test protocol, as created
by \code{\link{set_protocol_manual}} or \code{\link{get_protocol}}.}
}
\description{
\code{add_protocol()} adds a predefined test protocol to an existing set of
data from an exercise test.
}
