% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{add_restart}
\alias{add_restart}
\title{Spin image. Resulting file best viewed in Google Chrome.}
\usage{
add_restart(
  input,
  location = c(0.05, 0.95),
  fill = "white",
  color = "black",
  radius = 20,
  start_label = "Start",
  stop_label = "Stop",
  font_family = "inherit",
  font_size = 14,
  openfile = TRUE,
  output = input
)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{location}{a vector of coordinates for the start button}

\item{fill}{button fill color}

\item{color}{button text color}

\item{radius}{button radius}

\item{start_label}{start button text}

\item{stop_label}{stop button text}

\item{font_family}{button text font family}

\item{font_size}{button text font size}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Spin image. Resulting file best viewed in Google Chrome.
}
\examples{
\dontrun{
library(spiro)
spiro(fixed_radius = 3,
      cycling_radius = 1,
      file = "image_spin.svg") \%>\%
  image_spin(rpm = 1)
  }
}
