% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{add_weight}
\alias{add_weight}
\title{Calculated additional variables related to body weight to a spiroergometric
data set}
\usage{
add_weight(data, weight = NULL)
}
\arguments{
\item{data}{A \code{data.frame} containing the exercise test data.}

\item{weight}{A numeric value to manually set the participant's body weight.}
}
\value{
A \code{data.frame} containing the data amplified by the additional
variables.
}
\description{
\code{add_weight()} amplifies existing spiroergometric data by calculation of
additional variables.
}
\details{
Based on the participant's body weight relative oxygen uptake (VO2_rel) and
carbon dioxide (VCO2_rel) output are calculated. \code{weight} can be
manually overridden, e.g. if the meta-data's value is incorrect or rounded.

For running protocols, running economy (RE) is calculated.
}
