% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{apply_protocol}
\alias{apply_protocol}
\title{Apply a test protocol to a exercise testing dataset}
\usage{
apply_protocol(data, protocol)
}
\arguments{
\item{data}{A \code{data.frame} containing the exercise testing data
interpolated to seconds.}

\item{protocol}{A \code{data.frame} containing the test protocol, as created
by \code{\link{spiro_protocol}} or \code{\link{guess_protocol}}}
}
\description{
\code{apply_protocol()} adds a predifined test protocol to an existing set of
data from an exercise test.
}
