% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{get_features}
\alias{get_features}
\title{Extract features from an exercise test protocol}
\usage{
get_features(protocol)
}
\arguments{
\item{protocol}{A \code{data.frame} containing the raw protocol as given by
\code{\link{get_protocol}}, \code{\link{set_protocol}} or
\code{\link{set_protocol_manual}}.}
}
\description{
\code{get_features()} adds characteristic features to the load steps of
an exercise testing protocol.
}
