% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get_path}
\alias{get_path}
\title{Get the file path for a data file}
\usage{
get_path(name)
}
\arguments{
\item{name}{A character string, giving the file name to look for.}
}
\value{
A character string containing the file path.
}
\description{
\code{get_path()} performs a serch for a data file.
}
\details{
Considering the input to be a regular expression, this function will search
one level above the current working directory to find matching files.

If no matching files are found, \code{get_path} takes the input as a direct
path to a file. If more than one match is found, an error will be displayed.
}
