% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{get_testtype}
\alias{get_testtype}
\title{Guess a exercise test type from a corresponding test protocol}
\usage{
get_testtype(protocol)
}
\arguments{
\item{protocol}{A \code{data.frame} containing the test protocol, as given by
\code{\link{protocol_features}}.}
}
\value{
A character, either \code{"incremental"}, \code{"ramp"},
\code{"constant"} or \code{"other"}.
}
\description{
\code{get_testtype()} guesses which type of testing protocol a exercise test
is.
}
