% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{getstepmeans}
\alias{getstepmeans}
\title{Get mean data values for a step of an exercise test}
\usage{
getstepmeans(step_number, data, interval = 30)
}
\arguments{
\item{step_number}{An integer giving the number of the step}

\item{data}{A \code{data.frame} of the class \code{exed_*} containing the
data of the test}

\item{interval}{An integer giving the length of the computational interval
in seconds. For calculation the last \code{interval} seconds of the step
will be averaged.}
}
\description{
code{getstepmeans()} returns the average data values for a single step of an
exercise test.
}
