% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{guess_device}
\alias{guess_device}
\title{Guess the device used for a cardiopulmonary measurement}
\usage{
guess_device(file)
}
\arguments{
\item{file}{A character string, giving the path of the data file.}
}
\value{
A character string specifying the guessed device.
}
\description{
\code{guess_device()} guesses the device type of a metabolic cart based on
the characteristics of a raw data file. To get information on supported
devices visit \code{\link{spiro_import}}.
}
