% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{guess_protocol}
\alias{guess_protocol}
\title{Guess a test protocol from a corresponding exercise testing dataset}
\usage{
guess_protocol(data)
}
\arguments{
\item{data}{A \code{data.frame} containing the exercise testing data. It is
highly recommend to parse non-interpolated breath-by-breath data.}
}
\value{
A \code{data.frame} containing the characteristics of the test
protocol.
}
\description{
\code{guess_protocol()} guesses the underlying test protocol based on given
load data.
}
