% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartrate.R
\name{hr_add}
\alias{hr_add}
\alias{hr_import}
\alias{hr_apply}
\title{Import and apply heart rate data to a spiroergometric data file}
\usage{
hr_add(data, hr_file, hr_offset = 0)

hr_import(hr_file)

hr_apply(data, hr_data, hr_offset = 0)
}
\arguments{
\item{data}{A \code{data.frame} containing spiroergometric data interpolated
to every second.}

\item{hr_file}{The absolute or relative path of a \code{*tcx} file that
contains additional heart rate data.}

\item{hr_offset}{An integer, corresponding to the temporal offset of the
heart-rate file. By default the start of the heart rate measurement is
linked to the start of the gas exchange measurement.}

\item{hr_data}{A data.frame containing the imported heart rate data.}
}
\value{
A \code{data.frame} containing the spiroergometric and heart rate
data.
}
\description{
\code{hr_add()} imports an additional file containing heart rate data and
adds it to an existing spiroergometric data file.
}
\details{
Heart rate data will be imported from a \code{.tcx} file with
\code{hr_import()}. After interpolating the data to full seconds, it is then
applied to a given data set by \code{hr_apply()}.
}
\examples{
# Get example data
oxy_file <- spiro_example("zan_ramp")
hr_file <- spiro_example("hr_ramp.tcx")

# Import and process spiro data
oxy_data <- spiro(oxy_file)

# Add heart rate data
hr_add(oxy_data, hr_file)

}
