% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{image_animate}
\alias{image_animate}
\title{Animate image}
\usage{
image_animate(input, attribute = "opacity", values = c(0, 1, 0),
  duration = 10, openfile = TRUE, output = input)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{attribute}{Name of attribute. opacity is the default}

\item{values}{Vector of values to go between. Defaults to c(0,1,0)}

\item{duration}{Number of seconds for animation to last, defaults to 10 seconds}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Animate image
}
\examples{
library(spiro)
library(magrittr)
spiro(fixed_radius = 3, cycling_radius = 1)
}
