% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{image_merge}
\alias{image_merge}
\title{Merge images}
\usage{
image_merge(input, output = NA, copies = 1, delete_input = FALSE,
  openfile = TRUE)
}
\arguments{
\item{input}{Vector of file names to merge}

\item{output}{File name of .svg file to output.}

\item{copies}{The number of copies of the input file to make}

\item{delete_input}{Delete input file(s)}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE}
}
\value{
output name
}
\description{
Merge images
}
\examples{
library(spiro)
library(magrittr)
c(spiro(fixed_radius = 3, cycling_radius = 1),
  spiro(fixed_radius = 3, cycling_radius = 1,
        rotation = pi, colors = "firebrick")) \%>\%
  image_merge(output = "merged.svg")
}
