% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{image_rotate}
\alias{image_rotate}
\title{Rotate image.}
\usage{
image_rotate(
  input,
  degrees = 90,
  radians = NA,
  openfile = TRUE,
  output = input
)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{degrees}{Degrees to rotate image. Default is 90.}

\item{radians}{radians to rotate image. If specified, radians will override degrees}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
`image_rotate`
}
\examples{
library(spiro)
spiro(fixed_radius = 3,
      cycling_radius = 1,
      file = "image_rotate.svg") \%>\%
    image_rotate(degrees = 90)
}
