% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{image_scale}
\alias{image_scale}
\title{Scale image}
\usage{
image_scale(input, scale = 1, openfile = TRUE, output = input)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{scale}{Ratio by which to scale the image}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Scale image
}
\examples{
\dontrun{
library(spiro)
spiro(fixed_radius = 3,
      cycling_radius = 1,
      file = "image_scale") \%>\%
  image_scale(scale = 0.5)
  }
}
