% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{merge_image}
\alias{merge_image}
\title{Merge images}
\usage{
merge_image(input, output = NA, copies = 1, openfile = TRUE)
}
\arguments{
\item{input}{Vector of file names to merge}

\item{output}{File name of .svg file to output.}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE}
}
\value{
output name
}
\description{
Merge images
}
\examples{
library(spiro)
library(magrittr)
c(spiro(fixed_radius = 3, cycling_radius = 1),
  spiro(fixed_radius = 3, cycling_radius = 1,
        rotation = pi, colors = "firebrick")) \%>\%
  merge_image(output = "merged.svg")
}
