% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{pause_onload}
\alias{pause_onload}
\title{Pause all animations on loading}
\usage{
pause_onload(
  input,
  onload = "pauseAnimations();",
  openfile = TRUE,
  output = input
)
}
\arguments{
\item{input}{file name}

\item{onload}{javascript command to run onload.}

\item{openfile}{Open file in default program for .svg format. Defaults to TRUE}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Pause all animations on loading
}
