% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.spiro}
\alias{print.spiro}
\title{Printing spiro data frames}
\usage{
\method{print}{spiro}(x, digits = 2, ...)
}
\arguments{
\item{x}{A \code{data.frame} of the class \code{spiro} to be printed.}

\item{digits}{An integer giving the number of decimals to be rounded to.}

\item{...}{Passing of additional arguments to \code{print.data.frame()}.}
}
\description{
Printing method for spiro objects that rounds output to two decimals.
}
