% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{process_protocol}
\alias{process_protocol}
\title{Add information to an exercise test protocol}
\usage{
process_protocol(protocol, testtype = NULL)
}
\arguments{
\item{protocol}{A \code{data.frame} containing the raw protocol as given by
\code{\link{get_protocol}} or \code{\link{set_protocol_manual}}.}

\item{testtype}{A character, either \code{"ramp"}, \code{"constant"},
\code{"incremental"} or \code{"other"} for manually setting the test type.}
}
\description{
\code{process_protocol()} wraps protocol_features() and get_testtype() for
}
