% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{protocol_features}
\alias{protocol_features}
\title{Extract features from a test protocol}
\usage{
protocol_features(protocol)
}
\arguments{
\item{protocol}{A \code{data.frame} containing the raw protocol as given by
\code{\link{get_protocol}} or \code{\link{set_protocol_manual}}.}
}
\description{
\code{protocol_features()} adds characteristic features to the load steps of
an exercise testing protocol.
}
