% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{rotate_image}
\alias{rotate_image}
\title{Rotate image.}
\usage{
rotate_image(input, degrees = 90, radians = NA, openfile = TRUE,
  output = input)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{degrees}{Degrees to rotate image. Default is 90.}

\item{radians}{radians to rotate image. If specified, radians will override degrees}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
`rotate_image`
}
\examples{
library(spiro)
library(magrittr)
spiro(fixed_radius = 3, cycling_radius = 1) \%>\%
    rotate_image(degrees = 90)
}
