% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{safe_read}
\alias{safe_read}
\title{Reads file with a delay to reduce chance of file conflicts}
\arguments{
\item{file}{Name of file}
}
\value{
xml for use by xml2 functions
}
\description{
Reads file with a delay to reduce chance of file conflicts
}
\keyword{internal}
