% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{safe_write}
\alias{safe_write}
\title{Writes file with a delay to reduce chance of file conflicts}
\arguments{
\item{file}{Name of file}
}
\description{
Writes file with a delay to reduce chance of file conflicts
}
\keyword{internal}
