% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{scale_image}
\alias{scale_image}
\title{Scale image}
\usage{
scale_image(input, scale = 1, openfile = TRUE, output = input)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{scale}{Ratio by which to scale the image}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Scale image
}
\examples{
library(spiro)
library(magrittr)
spiro(fixed_radius = 3, cycling_radius = 1) \%>\%
    scale_image(scale = 0.5)
}
