% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{set_protocol_manual}
\alias{set_protocol_manual}
\alias{set_protocol_manual.default}
\alias{set_protocol_manual.data.frame}
\title{Manually setting a testing profile}
\usage{
set_protocol_manual(duration, load = NULL)

\method{set_protocol_manual}{default}(duration, load)

\method{set_protocol_manual}{data.frame}(data)
}
\description{
\code{set_protocol_manually()} allows to set any user-defined load profile
for an exercise test.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method when duration and load are
given separately

\item \code{data.frame}: Method for data.frames with duration and load
column
}}

