% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{shift_image}
\alias{shift_image}
\title{shift image}
\usage{
shift_image(input, x = 0, y = 0, openfile = TRUE, output = input)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{x}{Shift x. Can be positive or negative.}

\item{y}{Shift y. Can be positive or negative.
#' @param openfile Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
shift image
}
\examples{
library(spiro)
library(magrittr)
spiro(fixed_radius = 3, cycling_radius = 1) \%>\%
    shift_image(scale = 0.5)
}
