% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{spin_image}
\alias{spin_image}
\title{Spin image}
\usage{
spin_image(input, rpm = 1, rotation_point = c(0.5, 0.5),
  add_dot = FALSE, dot_color = "red", dot_size = 6,
  openfile = TRUE, output = input)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{rpm}{Number of rotations per minute. Positive numbers spin right. Negative numbers spin left. 0 is stationary.}

\item{rotation_point}{A vector of 2 numbers ranging from 0 to 1. Default is c(0.5, 0.5)}

\item{add_dot}{Adds a dot at the rotation point}

\item{dot_color}{Color of dot at center}

\item{dot_size}{Size of dot}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Spin image
}
\examples{
library(spiro)
library(magrittr)
spiro(fixed_radius = 3, cycling_radius = 1) \%>\%
    spin_image(rpm = 1)
}
