% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro-package.R
\docType{package}
\name{spiro-package}
\alias{spiro-package}
\title{spiro: Make Animated Spirographs}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Familiar to those who have spent hours with the classic toy "Spirograph," hypotrochoids and epitrochoids are shapes formed by tracing the path of a pen inside a circle as it rolls around inside or outside another circle. With this package, you can create, modify, and animate spirograph-like shapes in the svg format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjschne/spiro}
  \item \url{https://wjschne.github.io/spiro/}
  \item Report bugs at \url{https://github.com/wjschne/spiro/issues}
}

}
\author{
\strong{Maintainer}: W. Joel Schneider \email{w.joel.schneider@gmail.com} (\href{https://orcid.org/0000-0002-8393-5316}{ORCID})

Authors:
\itemize{
  \item W. Joel Schneider \email{w.joel.schneider@gmail.com} (\href{https://orcid.org/0000-0002-8393-5316}{ORCID})
}

}
\keyword{internal}
