% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro.R
\name{spiro}
\alias{spiro}
\title{Process raw data from metabolic carts/spiroergometric measures}
\usage{
spiro(
  file,
  device = NULL,
  weight = NULL,
  hr_file = NULL,
  hr_offset = 0,
  protocol = NULL
)
}
\arguments{
\item{file}{the name of the file which contains the spiroergometry data. The
file  will be found by partial matching of the regular expression within
the working directory (and one level above). Alternatively, \code{file} can
be an absolute or relative path.}

\item{device}{A character string, specifying the device for measurement. By
default the device type is guessed by the characteristics of \code{file}.
This can be overridden by setting the argument to \code{"zan"} or
\code{"cosmed"}.}

\item{weight}{Numeric value for participant's body weight, if the default
value saved in \code{file} should be overwritten.}

\item{hr_file}{the name of a \code{*tcx} file which contains additional heart
rate data. The file  will be found by partial matching of the regular
expression within the working directory (and one level above).
Alternatively, \code{file} can be an absolute or relative path.}

\item{hr_offset}{An integer, corresponding to the temporal offset of the
heart-rate file. By default the start of the heart rate measurement is
linked to the first load in the test protocol (i.e. the end of the
pre-measure).}

\item{protocol}{A \code{data.frame} by \code{spiro_protocol} containing the
test protocol. This is automatically guessed by default.}
}
\value{
A \code{data.frame} of the class \code{spiro_*} with the interpolated
spirometric data, the corresponding load and (if supplied) additional heart
rate data.
The attribute \code{"protocol"} provides additional information
on the underlying testing protocol. The attribute \code{"info"} contains
additional meta-data from the original raw file.
}
\description{
\code{spiro} wraps multiple functions to import, process and amplify raw data
from metabolic carts into a \code{data.frame}.
}
\details{
This function performs multiple operations on raw data from metabolic carts.
It imports the raw data from a file, which might be complemented by an
additional \code{*tcx} file with heart rate data.

Breath-by-breath data is linearly interpolated to get data points for every
second. Per default, based on the given velocity data, the underlying testing
protocol is guessed and applied to the data. Additional variables of gas
exchange are calculated for further analysis.

Per default the body weight saved in the file's metadata is used for
calculations. It is possible to supply \code{weight} manually to the
function.
}
