% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{spiro_add}
\alias{spiro_add}
\title{Add additional calculated variables to a spiroergometrical data set}
\usage{
spiro_add(data, weight = NULL)
}
\arguments{
\item{data}{A \code{data.frame} containing the exercise test data.}

\item{weight}{A numeric value to manually set the participant's body weight.}
}
\value{
A \code{data.frame} containing the data amplified by the additional
variables.
}
\description{
\code{spiro_add()} amplifies existing spiroergometric data by calculation of
additional variables.
}
\details{
Based on the participant's body weight relative oxygen uptake (VO2_rel) and
carbon dioxide (VCO2_rel) output are calculated. \code{weight} can be
manually overridden, e.g. if the meta-data's value is incorrect or rounded.

For running protocols, running economy (RE) is calculated.

Carbohydrate and fat oxidation rates are calculated from gas exchange data
using the formula by Peronnet (1991).
}
