% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{spiro_example}
\alias{spiro_example}
\title{Get path to spiro example}
\usage{
spiro_example(file)
}
\arguments{
\item{file}{Name of the file.}
}
\description{
\code{spiro_example} return the file path for example data files within the
\code{spiro} package.
}
