% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{spiro_glance}
\alias{spiro_glance}
\alias{spiro_glance.default}
\alias{spiro_glance.spiro_clt}
\title{Overall summarise data from spiroergometric tests}
\usage{
spiro_glance(data, interval)

\method{spiro_glance}{default}(data, interval = 30)

\method{spiro_glance}{spiro_clt}(data, interval = 30)
}
\arguments{
\item{data}{A \code{data.frame} of the class \code{spiro_*}, as it is
generated by \code{spiro()}.}

\item{interval}{Numeric value, giving the length of the computational
interval in seconds.}
}
\description{
\code{spiro_glance} returns a \code{data.frame} overall summarising the
parameters of a spiroergometric test.
}
\details{
This function generates a overall summary of gas exchange and cardiac
parameters for spiroergometric tests.

For constant load tests, the mean parameters for the given load are
calculated over all steps. The length of the computational interval before
the end of each step is given by \code{interval}.

For other test protocols, the maximum parameters are calculated. Data is
smoothed before by using a rolling average over the length of
\code{interval}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: The default method, which gives maximal values of
the rolling averages.

\item \code{spiro_clt}: Method for constant load tests. Gives an average of
all steps performed at the constant load.
}}

