% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro_get.R
\name{spiro_import}
\alias{spiro_import}
\title{Import raw data from spiroergometric devices (deprecated)}
\usage{
spiro_import(file, device = NULL, anonymize = TRUE)
}
\arguments{
\item{file}{The absolute or relative path of the file that contains the gas
exchange data.}

\item{device}{A character string, specifying the device for measurement. By
default the device type is guessed by the characteristics of the
\code{file}. This can be overridden by setting the argument to
\code{"cortex"}, \code{"cosmed"}, \code{"vyntus"} or \code{"zan"}.}

\item{anonymize}{Whether meta data should be anonymized during import.
Defaults to TRUE. See \code{\link{get_anonid}} for more information.}
}
\description{
This function has been deprecated as of package version \code{0.2.0}. It will
be removed in the next version release. Please use \code{\link{spiro}} for
automated import and processing or \code{\link{spiro_raw}} to import only raw
data.
}
