% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{spiro_import_cortex}
\alias{spiro_import_cortex}
\title{Import raw data from Cortex spiroergometric devices}
\usage{
spiro_import_cortex(file)
}
\arguments{
\item{file}{A character string, giving the path of the data file.}
}
\value{
A \code{data.frame} with data. The attribute \code{info} contains
addition meta-data retrieved from the original file.
}
\description{
\code{spiro_import_cortex()} retrieves cardiopulmonary data from cortex
metabolic cart files.
}
