% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro_import.R
\name{spiro_import_cosmed}
\alias{spiro_import_cosmed}
\title{Import raw data from COSMED spiroergometric devices}
\usage{
spiro_import_cosmed(file)
}
\arguments{
\item{file}{A character string, giving the path of the data file.}
}
\value{
A \code{data.frame} with data. The attribute \code{info} contains
addition meta-data retrieved from the original file.
}
\description{
\code{spiro_import_cosmed()} retrieves cardiopulmonary data from ZAN
metabolic cart files.
}
