% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro_import.R
\name{spiro_import_zan}
\alias{spiro_import_zan}
\title{Import raw data from ZAN spiroergometric devices}
\usage{
spiro_import_zan(file)
}
\arguments{
\item{file}{A character string, giving the path of the data file.}
}
\value{
A \code{data.frame} with data. The attribute \code{info} contains
addition meta-data retrieved from the original file.
}
\description{
\code{spiro_import_zan()} retrieves cardiopulmonary data from ZAN
metabolic cart files.
}
