% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{spiro_max}
\alias{spiro_max}
\title{Return maximum values from spiroergometric tests}
\usage{
spiro_max(data, smooth = 30, hr_smooth = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} of the class \code{spiro}, as it is
generated by \code{\link{spiro}}.}

\item{smooth}{An integer, giving the length of the computational
interval for the rolling average.}

\item{hr_smooth}{A logical, whether smoothing should also apply to heart rate
data. Default is \code{FALSE}, which means that the absolute maximum heart rate
value is taken without smoothing.}
}
\description{
\code{spiro_max()} returns a \code{data.frame} with the maximum parameter
parameters of a spiroergometric test.
}
\details{
Before calculating the maximum values, data is smoothed by using a rolling
average over the length of \code{interval}.
}
\examples{
# Import and process example data sets
gxt_data <- spiro(file = spiro_example("zan_gxt"))

spiro_max(gxt_data)
}
