% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{spiro_name}
\alias{spiro_name}
\title{checks if file exists and creates unique name if not}
\arguments{
\item{file}{Name of file}
}
\value{
Name of file
}
\description{
checks if file exists and creates unique name if not
}
\keyword{internal}
