% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{spiro_plot}
\alias{spiro_plot}
\title{Plot data from cardiopulmonary exercise data files}
\usage{
spiro_plot(data, which = 1:9, smooth = 15)
}
\arguments{
\item{data}{A \code{data.frame} of the class \code{spiro_*}, as it is
generated by \code{\link{spiro}}.}

\item{which}{A numeric integer setting the plot panels to be displayed.}

\item{smooth}{Integer for calculation of a rolling average (in seconds)
for gas exchange measures.}
}
\description{
\code{spiro_plot_*()} returns a \code{ggplot2} graph visualising
data from cardiopulmonary exercise testing.
}
\details{
This function provides a shortcut for visualising spiroergometric data from
\code{\link{spiro}} with the help of \code{ggplot2}.
}
\examples{
# Import and process example data
ramp_data <- spiro(file = spiro_example("zan_ramp"),
                  hr_file = spiro_example("hr_ramp.tcx"))

spiro_plot(ramp_data)

}
