% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{spirograph}
\alias{spirograph}
\title{Make coordinates for spirograph}
\usage{
spirograph(fixed_radius = 3, cycling_radius = 1,
  pen_radius = cycling_radius, origin_x = 0, origin_y = 0,
  windings = cycling_radius, rotation = 0, start_angle = 0,
  npoints = round(abs(1000 * windings), 0), ngroups = 1)
}
\arguments{
\item{fixed_radius}{The radius of the fixed circle being cycled around. Can be positive or negative.}

\item{cycling_radius}{The radius of the moving circle cycling around the fixed circle. Can be positive or negative.}

\item{pen_radius}{The pen placement inside the moving circle cycling around the fixed circle. Can be positive or negative.}

\item{origin_x}{The x coordinate of the center of the fixed circle}

\item{origin_y}{The y coordinate of the center of the fixed circle}

\item{windings}{The number of times the the moving circle winds around the fixed circle.}

\item{rotation}{The number of radians by which the entire figure should be rotated around the origin of the fixed circle}

\item{start_angle}{The number of radians on the fixed circle's arc where the moving circle starts cycling.}

\item{npoints}{= The number of points to generate}

\item{ngroups}{= Number of groups in group_id variable}
}
\value{
A tibble with x and y coordinates of the spirograph, row id, and group id
}
\description{
Make coordinates for spirograph
}
\examples{
spirograph(fixed_radius = 3, cycling_radius = 1)
}
