% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{spiroSVG}
\alias{spiroSVG}
\title{Generate spirograph image in the svg format}
\usage{
spiroSVG(d, path = "spiro.svg",
  color = (scales::viridis_pal())(length(unique(d$group_id))),
  transparency = 1, rule = "evenodd", openfile = TRUE,
  aspect_ratio = 1)
}
\arguments{
\item{d}{A tibble output of the spirograph function}

\item{path}{Name of the output file. Defaults to "spiro.svg"}

\item{color}{Vector of colors for groups. Defaults to viridis scale.}

\item{transparency}{Transparency of fills. Ranges from 0 to 1. Defaults to 1.}

\item{rule}{Character value specifying the path fill mode: either "evenodd" or "winding". Defaults to "evenodd"}

\item{openfile}{Open file in default program for .svg format. Defaults to TRUE}

\item{aspect_ratio}{Aspect ratio of the image. Defaults to 1.}
}
\value{
A tibble with x and y coordinates of the spirograph, row id, and group id
}
\description{
Generate spirograph image in the svg format
}
\examples{
d <- spirograph(fixed_radius = 3, cycling_radius = 1)
spiroSVG(d)
}
