% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{string_bezier}
\alias{string_bezier}
\title{Make string bezier}
\usage{
string_bezier(file = "bezier.svg", x = c(0, 0, 1), y = c(1, 0, 0),
  n = 20, color = "royalblue", openfile = TRUE, ...)
}
\arguments{
\item{file}{Name of the output file. Defaults to "bezier.svg"}

\item{x}{The x coordinates of the 3 points of the curve}

\item{y}{The y coordinates of the 3 points of the curve}

\item{n}{The number of segments. Defaults to 20.}

\item{color}{Vector of colors}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{...}{parameters passed to the segments function}
}
\value{
file by default or tibble with data if savefile if FALSE
}
\description{
Make string bezier
}
\examples{
library(spiro)
library(magrittr)
string_bezier(
   file = "bezier_redblueblack.svg",
   x = c(0,0,1),
   y = c(1,0,0),
   color = c("red","blue","black"),
   n = 100,
   lwd = 0.3
   )
string_bezier(
 file = "bezier_sequence.svg",
 x = c(-1,1,0,-1,1,-1, 0,1,-1,1),
 y = c( 0,0,1, 0,0, 0,-1,0, 0,0),
 color = c("#006B6B","#FFFFFF","black","black","black"),
 n = 200,
 lwd = 5,
 ljoin = 1
 ) \%>\%
   add_background(color = "#108998")
}
