% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{to_seconds}
\alias{to_seconds}
\title{Convert time data to seconds}
\usage{
to_seconds(time_data)
}
\arguments{
\item{time_data}{A character vector containing the time data in the format
hh:mm:ss or mm:ss.}
}
\description{
\code{to_seconds()} converts time data of the form hh:mm:ss to seconds.
}
