% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weight.R
\name{add_weight}
\alias{add_weight}
\title{Calculate additional variables related to body weight for cardiopulmonary
exercise testing data}
\usage{
add_weight(data, weight = NULL)
}
\arguments{
\item{data}{A \code{data.frame} of the class \code{spiro} containing the gas
exchange data. Usually the output of a \code{\link{spiro}} call.}

\item{weight}{A numeric value to manually set the participant's body weight.}
}
\value{
A \code{data.frame} of the class \code{spiro} containing the
cardiopulmonary exercise testing data including variables relative to body
weight.
}
\description{
\code{add_weight()} amplifies existing spiroergometric data by the
calculation of body weight-related variables.
}
\details{
Based on the participant's body weight relative oxygen uptake (VO2_rel) and
carbon dioxide (VCO2_rel) output are calculated. \code{weight} will be
received from the original file's meta data, but can be manually overridden,
e.g. if the value in the meta data is incorrect or rounded.

For running protocols, running economy (RE) is calculated.
}
\examples{
# get example file
file <- spiro_example("zan_gxt")

s <- spiro(file)
add_weight(s, weight = 65.3)
}
