% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartrate.R
\name{hr_add}
\alias{hr_add}
\alias{hr_import}
\alias{hr_apply}
\title{Import and apply heart rate data to a spiroergometric data file}
\usage{
hr_add(data, hr_file, hr_offset = 0)

hr_import(hr_file)

hr_apply(data, hr_data, hr_offset = 0)
}
\arguments{
\item{data}{A \code{data.frame} containing spiroergometric data interpolated
to every second.}

\item{hr_file}{The name of a \code{*tcx} file which contains additional heart
rate data. The file  will be found by partial matching of the regular
expression within the working directory (and one level above).
Alternatively, \code{hr_file} can be an absolute or relative path.}

\item{hr_offset}{An integer, corresponding to the temporal offset of the
heart-rate file. By default the start of the heart rate measurement is
linked to the first load in the test protocol (i.e. the end of the
pre-measure).}

\item{hr_data}{A data.frame containing the imported heart rate data.}
}
\value{
A \code{data.frame} containing the spiroergometric and heart rate
data.
}
\description{
\code{hr_add()} imports an additional file containing heart rate data and
adds it to an existing spiroergometric data file.
}
\details{
Heart rate data will be imported from a \code{.tcx} file with
\code{hr_import()}. After interpolating the data to full seconds, it is then
applied to a given data set by \code{hr_apply()}.
}
