% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{image_animate}
\alias{image_animate}
\title{Animate image. Resulting file best viewed in Google Chrome.}
\usage{
image_animate(
  input,
  attribute = "opacity",
  values = c(0, 1, 0),
  duration = 10,
  openfile = TRUE,
  output = input
)
}
\arguments{
\item{input}{File name of .svg file to input}

\item{attribute}{Name of attribute. opacity is the default}

\item{values}{Vector of values to go between. Defaults to c(0,1,0)}

\item{duration}{Number of seconds for animation to last, defaults to 10 seconds}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE.}

\item{output}{File name of .svg file to output. Default is to overwrite the input file.}
}
\value{
output name
}
\description{
Animate image. Resulting file best viewed in Google Chrome.
}
\examples{
\dontshow{.old_wd <- setwd(tempdir())}
library(spiro)
spiro(fixed_radius = 3,
      cycling_radius = 1,
      file = "image_animate.svg") \%>\%
   image_animate(attribute = "opacity",
                 duration = 2,
                 values = c(0.3,0.9,0.3))
\dontshow{setwd(.old_wd)}
}
