% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{image_merge}
\alias{image_merge}
\title{Merge images}
\usage{
image_merge(
  input,
  output = NA,
  copies = 1,
  delete_input = TRUE,
  openfile = TRUE
)
}
\arguments{
\item{input}{Vector of file names to merge}

\item{output}{File name of .svg file to output.}

\item{copies}{The number of copies of the input file to make}

\item{delete_input}{Delete input file(s)}

\item{openfile}{Open file in default program for .svg format. Defaults to FALSE}
}
\value{
output name
}
\description{
Merge images
}
\examples{
\dontshow{.old_wd <- setwd(tempdir())}
library(spiro)
c(spiro(fixed_radius = 6, transparency = 0.5),
  spiro(fixed_radius = 3, transparency = 0.5)
) \%>\%
  image_merge(
    output = "image_merge.svg",
    delete_input = TRUE)
\dontshow{setwd(.old_wd)}
}
