% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{knit_print.spiro}
\alias{knit_print.spiro}
\title{define a knit_print method for objects of the class spiro. It inserts the .svg file into the document with the knitr::include_graphics function}
\usage{
knit_print.spiro(x, ...)
}
\arguments{
\item{x}{Name of the output file.}
}
\description{
define a knit_print method for objects of the class spiro. It inserts the .svg file into the document with the knitr::include_graphics function
}
\keyword{internal}
