% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{knit_print.spiro}
\alias{knit_print.spiro}
\title{define a knit_print method for objects of the class spiro}
\usage{
knit_print.spiro(file, ...)
}
\arguments{
\item{file}{Name of the output file.}

\item{...}{Parameters passed to knitr::include_graphics}
}
\description{
define a knit_print method for objects of the class spiro
}
