% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spirograph.R
\name{print.spiro}
\alias{print.spiro}
\title{Custom print function for spiro objects}
\usage{
\method{print}{spiro}(file)
}
\arguments{
\item{file}{Name of the output file.}
}
\description{
Custom print function for spiro objects
}
