% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro-package.R
\docType{package}
\name{spiro-package}
\alias{spiro-package}
\title{spiro: Manage Data from Cardiopulmonary Exercise Testing}
\description{
Import, process, summarize and visualize raw data from metabolic carts. See Robergs, Dwyer, and Astorino (2010) \doi{10.2165/11319670-000000000-00000} for more details on data processing.
}
\section{Main functions}{

\itemize{
\item Use \code{\link{spiro}} for the import and processing of
cardiopulmonary exercise testing data.
\item Use \code{\link{spiro_summary}} and \code{\link{spiro_max}} for
a step wise summary or the maximum parameter values of the
imported data.
\item Use \code{\link{spiro_plot}} to visualize the data.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/spiro}
  \item \url{https://docs.ropensci.org/spiro/}
  \item Report bugs at \url{https://github.com/ropensci/spiro/issues}
}

}
\author{
\strong{Maintainer}: Simon Nolte \email{s.nolte@dshs-koeln.de} (\href{https://orcid.org/0000-0003-1643-1860}{ORCID})

Other contributors:
\itemize{
  \item Manuel Ramon (reviewed the package (v. 0.0.5) for rOpenSci, see <https://github.com/ropensci/software-review/issues/541>) [reviewer]
  \item James Hunter (reviewed the package (v. 0.0.5) for rOpenSci, see <https://github.com/ropensci/software-review/issues/541>) [reviewer]
}

}
\keyword{internal}
