% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro-package.R
\docType{package}
\name{spiro-package}
\alias{spiro-package}
\title{spiro: Manage Data from Cardiopulmonary Exercise Testing}
\description{
The spiro package enables a fast workflow with raw files from metabolic
carts. It provides simple tools for the import, processing, summary and
visualization of cardiopulmonary exercise test data.
}
\section{Main functions}{

\itemize{
\item Use \code{\link{spiro}} for the import and processing of
cardiopulmonary exercise testing data.
\item Use \code{\link{spiro_summary}} and \code{\link{spiro_max}} for
a step wise summary or the maximum parameter values of the
imported data.
\item Use \code{\link{spiro_plot}} to visualize the data.
}
}

