% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro-package.R
\docType{package}
\name{spiro-package}
\alias{spiro-package}
\title{spiro: Work with data from cardiopulmonary exercise testing}
\description{
The spiro package enables a fast workflow with raw files from metabolic
carts/spiroergometry systems. It provides simple tools for data import,
processing, summary and visualisation.
}
\section{Main functions}{

\itemize{
\item Use \code{\link{spiro}} for import and processing of cardiopulmonary
exercise testing data.
\item Use \code{\link{spiro_summary}} and \code{\link{spiro_max}} for
a stepwise summary or the maximum parameter values of the
imported data.
\item Use \code{\link{spiro_plot}} to visualize the data.
}
}

