% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro-package.R
\docType{package}
\name{spiro-package}
\alias{spiro-package}
\title{spiro: A package to work with data from spiroergometric testing}
\description{
The spiro package enables a fast workflow with raw files from metabolic
carts/spiroergometry systems. It provides simple tools for data import,
processing, summary and visualisation.
}
\section{Main functions}{

\itemize{
\item Use \code{\link{spiro}} for one-step data import and processing from
raw metabolic cart files.
\item Use \code{\link{spiro_summary}} and \code{\link{spiro_max}} for
getting a stepwise summary or the maximum parameter values of the
imported data.
\item Use the \code{\link{spiro_plot}} functions as a shortcut for
visualising the imported data.
}
}

