% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro-package.R, R/spirograph.R
\name{spiro}
\alias{spiro}
\title{spiro: Creates spirographs
#' @docType package}
\usage{
spiro(fixed_radius = 3, cycling_radius = 1,
  pen_radius = cycling_radius, windings = cycling_radius,
  color_groups = 1, colors = NA, transparency = NA,
  color_cycles = 1, color_sort = FALSE, draw_fills = TRUE,
  line_width = 1, origin_x = 0, origin_y = 0, xlim = NULL,
  ylim = NULL, rotation = 0, start_angle = 0,
  points_per_polygon = round(abs(1000 * windings), 0),
  rule = c("evenodd", "winding"), openfile = TRUE,
  end_at_beginning = F, savefile = TRUE, file = NA, ...)
}
\arguments{
\item{fixed_radius}{The radius of the fixed circle being cycled around. Can be positive or negative. Defaults to 3.}

\item{cycling_radius}{The radius of the moving circle cycling around the fixed circle. Can be positive or negative. Defaults to 1.}

\item{pen_radius}{The pen placement inside the moving circle cycling around the fixed circle. Can be positive or negative.}

\item{windings}{The number of times the the moving circle winds around the fixed circle.}

\item{color_groups}{= Number of groups in group_id variable}

\item{colors}{Vector of colors for groups. Defaults to black when color_groups is 1 and to the viridis scale when color_groups > 1.}

\item{transparency}{Transparency of colors. Ranges from 0 to 1. Default is NA (i.e., leave the colors as specified in the colors argument).}

\item{color_cycles}{The number of times the color vector  is recycled}

\item{color_sort}{Arranges all paths of the same color to be written in groups}

\item{draw_fills}{Draw fills. Default is TRUE}

\item{line_width}{The width of lines. Only relevant when draw_fills is FALSE.}

\item{origin_x}{The x coordinate of the center of the fixed circle}

\item{origin_y}{The y coordinate of the center of the fixed circle}

\item{xlim}{The limits of the x coordinates. For example, c(-5,5). NULL means that the limits are determined by the data point that has the largest distance from the origin.}

\item{ylim}{The limits of the y coordinates. For example, c(-5,5). NULL means that the limits are determined by the data point that has the largest distance from the origin.}

\item{rotation}{The number of radians by which the entire figure should be rotated around the origin of the fixed circle}

\item{start_angle}{The number of radians on the fixed circle's arc where the moving circle starts cycling.}

\item{points_per_polygon}{= The number of points for each color grouping}

\item{rule}{Character value specifying the path fill mode: either "evenodd" or "winding". Defaults to "evenodd"}

\item{openfile}{Open file in default program for .svg format. Defaults to TRUE.}

\item{end_at_beginning}{Adds the first point to the end of the figure but with the colors of the last color group. Defaults to FALSE.}

\item{savefile}{Save a file. Defaults to TRUE.}

\item{file}{Name of the output file. Defaults to "spiro.svg"}

\item{...}{parameters passed to the par function}
}
\value{
file by default or tibble with data if savefile if FALSE
}
\description{
The spiro package creates spirographs in the svg format
and can create svg animations of spirographs.'
}
\examples{
spiro(fixed_radius = 3, cycling_radius = 1)
}
\author{
W. Joel Schneider
#'
}
