% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiro_example.R
\name{spiro_example}
\alias{spiro_example}
\title{Get path to spiro example}
\usage{
spiro_example(file = NULL)
}
\arguments{
\item{file}{Name of the file, either "zan_gxt", "zan_ramp" or "hr_ramp.tcx".
Leave the argument empty to get a vector with the paths of all three
example files. Note that the package contains more example data, which is
used for internal testing.}
}
\value{
A character vector with the absolute file path of the example
file(s).
}
\description{
\code{spiro_example} returns the file path for example data files within the
\code{spiro} package.
}
\examples{
# get path of a specific example data file
spiro_example("zan_gxt")

# get all paths of example data files
spiro_example()
}
