% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{spiro_max}
\alias{spiro_max}
\title{Return maximum values from spiroergometric tests}
\usage{
spiro_max(data, interval = 30)
}
\arguments{
\item{data}{A \code{data.frame} of the class \code{spiro}, as it is
generated by \code{\link{spiro}}.}

\item{interval}{An integer, giving the length of the computational
interval in seconds.}
}
\description{
\code{spiro_max()} returns a \code{data.frame} with the maximum parameter
parameters of a spiroergometric test.
}
\details{
Before calculating the maximum values, data is smoothed by using a rolling
average over the length of \code{interval}.
}
\examples{
# Import and process example data sets
gxt_data <- spiro(file = spiro_example("zan_gxt"))

spiro_max(gxt_data)

}
