% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{spiro_plot}
\alias{spiro_plot}
\alias{spiro_plot_VO2}
\alias{spiro_plot_HR}
\title{Plot data from spiroergometry files}
\usage{
spiro_plot_VO2(data, smooth = 15)

spiro_plot_HR(data)
}
\arguments{
\item{data}{A \code{data.frame} of the class \code{spiro_*}, as it is
generated by \code{\link{spiro}}.}

\item{smooth}{Integer for calculation of a rolling average (in seconds)
for gas exchange measures.}
}
\description{
\code{spiro_plot_*()} returns a \code{ggplot2} graph visualising
data from a spiroergometric test.
}
\details{
This function provides a shortcut for visualising spiroergometric data from
\code{\link{spiro}} with \code{ggplot2}. It generates a graph showing the
(smoothed) interpolated data, as well as the underlying test protocol.
}
\examples{
# Import and process example data
ramp_data <- spiro(file = spiro_example("zan_ramp"),
                  hr_file = spiro_example("hr_ramp.tcx"))

spiro_plot_VO2(ramp_data)

spiro_plot_HR(ramp_data)

}
