% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{spiro_plot}
\alias{spiro_plot}
\title{Plot data from cardiopulmonary exercise data files}
\usage{
spiro_plot(data, which = 1:9, smooth = 15, base_size = 13, ...)
}
\arguments{
\item{data}{A \code{data.frame} of the class \code{spiro}, as it is
generated by \code{\link{spiro}}.}

\item{which}{A numeric integer setting the plot panels to be displayed. The
panels are numbered in the order of the traditional Wasserman 9-Panel
Plot:
\itemize{
\item 1: VE over time
\item 2: HR and oxygen pulse over time
\item 3: VO2, VCO2 and load over time
\item 4: VE over VCO2
\item 5: V-Slope: HR and VCO2 over VO2
\item 6: EQVO2 and EQVCO2 over time
\item 7: VT over VE
\item 8: RER over time
\item 9: PetO2 and PetCO2 over time
}}

\item{smooth}{Integer for calculation of a rolling average (in seconds)
for gas exchange measures.}

\item{base_size}{An integer controlling the base size of the plots (in pts).}

\item{...}{Arguments passed to ggplot2::theme() to modify the appearance of
the plots.}
}
\description{
\code{spiro_plot()} returns a \code{ggplot2} graph visualizing
data from cardiopulmonary exercise testing.
}
\details{
This function provides a shortcut for visualizing spiroergometric data from
\code{\link{spiro}} with the help of \code{ggplot2}.
}
\examples{
# Import and process example data
ramp_data <- spiro(
  file = spiro_example("zan_ramp"),
  hr_file = spiro_example("hr_ramp.tcx")
)

# Diplay the traditional Wasserman 9-Panel Plot
spiro_plot(ramp_data)

# Display selected panels, here V-Slope
spiro_plot(ramp_data, which = 5)
}
