% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{spiro_summary}
\alias{spiro_summary}
\title{Stepwise summarise data from spiroergometric tests}
\usage{
spiro_summary(data, interval = 30)
}
\arguments{
\item{data}{A \code{data.frame} of the class \code{spiro_*}, as it is
generated by \code{\link{spiro}}.}

\item{interval}{An integer giving the length of the computational
interval in seconds.}
}
\description{
\code{spiro_summary()} returns a \code{data.frame} summarising the main
parameters for each step of a spiroergometric test.
}
\details{
This function generates mean values of gas exchange and cardiac parameters
for all steps of a spiroergometric test. For calculation the given
\code{interval} before the end of each step is used.
}
